% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_pvalues.R
\name{asymptotic_pvalues}
\alias{asymptotic_pvalues}
\title{This function finds the p values of several tests based on large sample theory}
\usage{
asymptotic_pvalues(x, n, m)
}
\arguments{
\item{x}{a vector of test statistics}

\item{n}{size of sample 1}

\item{m}{size of sample 2}
}
\value{
A vector of p values.
}
\description{
This function finds the p values of several tests based on large sample theory
}
