% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bincounter}
\alias{bincounter}
\title{find counts in bins. Useful for power calculations. Replaces hist command from R.}
\usage{
bincounter(x, bins)
}
\arguments{
\item{x}{numeric vector}

\item{bins}{numeric vector}
}
\value{
Integer vector of counts
}
\description{
find counts in bins. Useful for power calculations. Replaces hist command from R.
}
\keyword{internal}
