% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_power.R
\name{chi_power}
\alias{chi_power}
\title{This function runs the chi-square test for continuous or discrete data}
\usage{
chi_power(
  rxy,
  alpha = 0.05,
  B = 1000,
  xparam,
  yparam,
  nbins = c(50, 10),
  minexpcount = 5,
  typeTS
)
}
\arguments{
\item{rxy}{a function to generate data}

\item{alpha}{=0.05 type I error probability of test}

\item{B}{=1000 number of simulation runs}

\item{xparam}{vector of parameter values}

\item{yparam}{vector of parameter values}

\item{nbins}{=c(50, 10) number of desired bins}

\item{minexpcount}{=5 smallest number of counts required in each bin}

\item{typeTS}{type of problem, continuous/discrete, with/without weights}
}
\value{
A matrix of power values
}
\description{
This function runs the chi-square test for continuous or discrete data
}
