% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_cont_noweights}
\alias{gen_cont_noweights}
\title{simulate continuous data without weights}
\usage{
gen_cont_noweights(x, y, TSextra)
}
\arguments{
\item{x}{first data set}

\item{y}{second data set}

\item{TSextra}{extra stuff}
}
\value{
A list of permuted vectors
}
\description{
simulate continuous data without weights
}
\keyword{internal}
