% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_cont_weights}
\alias{gen_cont_weights}
\title{simulate continuous data with weights}
\usage{
gen_cont_weights(x, y, wx, wy, TSextra)
}
\arguments{
\item{x}{first data set}

\item{y}{second data set}

\item{wx}{weights of first data set}

\item{wy}{weights of second data set}

\item{TSextra}{extra stuff}
}
\value{
A list of permuted vectors
}
\description{
simulate continuous data with weights
}
\keyword{internal}
