% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_disc}
\alias{gen_disc}
\title{simulate new discrete data}
\usage{
gen_disc(dtax, dtay, vals, TSextra)
}
\arguments{
\item{dtax}{first data set, counts}

\item{dtay}{second data set, counts}

\item{vals}{values of discrete random variable}

\item{TSextra}{extra stuff}
}
\value{
A list of permuted vectors
}
\description{
simulate new discrete data
}
\keyword{internal}
