% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_sim_data}
\alias{gen_sim_data}
\title{simulate continuous data without weights}
\usage{
gen_sim_data(dta, TSextra)
}
\arguments{
\item{dta}{data set}

\item{TSextra}{extra stuff}
}
\value{
A list of permuted vectors
}
\description{
simulate continuous data without weights
}
\keyword{internal}
