% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{myTS2}
\alias{myTS2}
\title{a local function needed for the vignette}
\usage{
myTS2(x, y, vals)
}
\arguments{
\item{x}{An integer vector.}

\item{y}{An integer vector.}

\item{vals}{A numeric vector with the values of the discrete rv.}
}
\value{
A vector with test statistics
}
\description{
a local function needed for the vignette
}
