% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perm_test_cont}
\alias{perm_test_cont}
\title{run permutation test.}
\usage{
perm_test_cont(x, y, TS, typeTS, TSextra, wx, wy, B = 5000L)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}

\item{TS}{routine to calculate test statistics for non-chi-square tests}

\item{typeTS}{type of a test statistic}

\item{TSextra}{additional info passed to TS, if necessary}

\item{wx}{a numeric vector of weights of x.}

\item{wy}{a numeric vector of weights of y.}

\item{B}{=5000, number of simulation runs.}
}
\value{
A list with test statistics and p values
}
\description{
run permutation test.
}
\keyword{internal}
