% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perm_test_disc}
\alias{perm_test_disc}
\title{run permutation test.}
\usage{
perm_test_disc(x, y, vals, TS, typeTS, TSextra, samplingmethod = 1L, B = 5000L)
}
\arguments{
\item{x}{A vector of counts or weights.}

\item{y}{A vector of counts or weights.}

\item{vals}{A numeric vector. Indicates discrete data.}

\item{TS}{routine to calculate test statistics for non-chi-square tests}

\item{typeTS}{type of a test statistic}

\item{TSextra}{additional info passed to TS, if necessary}

\item{samplingmethod}{=1, 1 for independence sampling, 2 for MCMC}

\item{B}{Number of simulation runs.}
}
\value{
A list with test statistics and p values
}
\description{
run permutation test.
}
\keyword{internal}
