% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permute_disc}
\alias{permute_disc}
\title{permute discrete data}
\usage{
permute_disc(dta, samplingmethod = 1L)
}
\arguments{
\item{dta}{A list of numeric vectors.}

\item{samplingmethod}{=2 , 1 for independence sampling by expansion or 2 for MCMC}
}
\value{
A list of permuted vectors
}
\description{
permute discrete data
}
\keyword{internal}
