% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{power_cont}
\alias{power_cont}
\title{Find the power of various discrete tests via permutation.}
\usage{
power_cont(
  rxy,
  TS,
  typeTS,
  TSextra,
  alpha = 0.05,
  B = as.numeric(c(1000, 1000)),
  xparam = 0,
  yparam = 0
)
}
\arguments{
\item{rxy}{a function that generates x and y data.}

\item{TS}{routine to calculate test statistics for non-chi-square tests}

\item{typeTS}{indicator for type of test statistics}

\item{TSextra}{additional info passed to TS, if necessary}

\item{alpha}{A numeric constant}

\item{B}{=c(1000, 1000) number of simulation runs for power and for p values}

\item{xparam}{arguments for r1.}

\item{yparam}{arguments for r2.}
}
\value{
A numeric matrix of powers
}
\description{
Find the power of various discrete tests via permutation.
}
\keyword{internal}
