% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_cont_LS.R
\name{power_cont_LS}
\alias{power_cont_LS}
\title{Find the power of various discrete tests via permutation.}
\usage{
power_cont_LS(rxy, alpha = 0.05, B = 1000, xparam = 0, yparam = 0)
}
\arguments{
\item{rxy}{a function that generates x and y data.}

\item{alpha}{A numeric constant}

\item{B}{Number of simulation runs.}

\item{xparam}{arguments for r1.}

\item{yparam}{arguments for r2.}
}
\value{
A numeric matrix of powers
}
\description{
Find the power of various discrete tests via permutation.
}
\keyword{internal}
