% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_cont_R.R
\name{power_cont_R}
\alias{power_cont_R}
\title{Find the power of built-in continuous two sample tests using Rcpp and parallel computing.}
\usage{
power_cont_R(
  rxy,
  xparam,
  yparam,
  TS,
  typeTS,
  TSextra,
  alpha = 0.05,
  B = 1000,
  maxProcessor
)
}
\arguments{
\item{rxy}{function to generate a list with data sets x, y and (optional) vals, weights}

\item{xparam}{first argument passed to rxy}

\item{yparam}{second argument passed to rxy}

\item{TS}{test statistic}

\item{typeTS}{which format has TS?}

\item{TSextra}{list of items passed TS}

\item{alpha}{=0.05, the level of the hypothesis test}

\item{B}{= 1000 number of simulation runs}

\item{maxProcessor}{maximum number of cores to use. If maxProcessor=1 no parallel computing is used.}
}
\value{
A numeric vector of power values.
}
\description{
Find the power of built-in continuous two sample tests using Rcpp and parallel computing.
}
