% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{power_disc}
\alias{power_disc}
\title{Find the power of various continuous tests via permutation.}
\usage{
power_disc(
  rxy,
  xparam,
  yparam,
  TS,
  typeTS,
  TSextra,
  samplingmethod = 1L,
  B = 1000L
)
}
\arguments{
\item{rxy}{a function that generates x and y data.}

\item{xparam}{arguments for r1.}

\item{yparam}{arguments for r2.}

\item{TS}{routine to calculate test statistics for non-chi-square tests}

\item{typeTS}{indicator for type of test statistics}

\item{TSextra}{additional info passed to TS, if necessary}

\item{samplingmethod}{=1, 1 for independence sampling, 2 for MCMC}

\item{B}{=1000 number of simulation runs}
}
\value{
A list values of test statistics
}
\description{
Find the power of various continuous tests via permutation.
}
\keyword{internal}
