% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerR.R
\name{powerR}
\alias{powerR}
\title{Find the power of two sample tests using Rcpp and parallel computing.}
\usage{
powerR(
  rxy,
  xparam,
  yparam,
  TS,
  typeTS,
  TSextra,
  alpha = 0.05,
  B = 1000,
  SuppressMessages,
  maxProcessor
)
}
\arguments{
\item{rxy}{function to generate a list with data sets x, y and (optional) vals, weights}

\item{xparam}{first argument passed to rxy}

\item{yparam}{second argument passed to rxy}

\item{TS}{test statistic}

\item{typeTS}{which format has TS?}

\item{TSextra}{list of items passed TS}

\item{alpha}{=0.05, the level of the hypothesis test}

\item{B}{= 1000 number of simulation runs}

\item{SuppressMessages}{= FALSE print informative messages?}

\item{maxProcessor}{maximum number of cores to use. If maxProcessor=1 no parallel computing is used.}
}
\value{
A numeric vector of power values.
}
\description{
Find the power of two sample tests using Rcpp and parallel computing.
}
