% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2sample-package.R
\docType{package}
\name{R2sample-package}
\alias{R2sample}
\alias{R2sample-package}
\title{R2sample: Various Methods for the Two Sample Problem}
\description{
The routine twosample_test() in this package runs the two sample test using various test statistic. The p values are found via permutation or large sample theory. The routine twosample_power() allows the calculation of the power in various cases, and plot_power() draws the corresponding power graphs. The routine run.studies allows a user to quickly study the power of a new method and how it compares to some of the standard ones.
}
\author{
\strong{Maintainer}: Wolfgang Rolke \email{wolfgang.rolke@upr.edu} (\href{https://orcid.org/0000-0002-3514-726X}{ORCID})

}
\keyword{internal}
