% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repC}
\alias{repC}
\title{cpp version of R routine rep}
\usage{
repC(x, times)
}
\arguments{
\item{x}{numeric vector}

\item{times}{integer vector}
}
\value{
A numeric vector
}
\description{
cpp version of R routine rep
}
\keyword{internal}
