% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_shiny}
\alias{run_shiny}
\title{Runs the shiny app associated with R2sample package}
\usage{
run_shiny()
}
\value{
No return value, called for side effect of opening a shiny app
}
\description{
Runs the shiny app associated with R2sample package
}
