% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{testC}
\alias{testC}
\title{run test using either simulation or permutation.}
\usage{
testC(dta, TS, typeTS, TSextra, B = 5000L)
}
\arguments{
\item{dta}{a list with the data}

\item{TS}{routine to calculate test statistics for non-chi-square tests}

\item{typeTS}{type of a test statistic}

\item{TSextra}{additional info passed to TS, if necessary}

\item{B}{=5000, number of simulation runs.}
}
\value{
A list with test statistics and p values
}
\description{
run test using either simulation or permutation.
}
\keyword{internal}
