% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ts_C}
\alias{ts_C}
\title{This function calculates the test statistics for continuous data}
\usage{
ts_C(typeTS, x, y, TS, TSextra, wx, wy)
}
\arguments{
\item{typeTS}{format of TS}

\item{x}{continuous data set}

\item{y}{continuous data set}

\item{TS}{routine}

\item{TSextra}{list passed to TS function}

\item{wx}{weights for x}

\item{wy}{weights for y}
}
\value{
A vector of numbers
}
\description{
This function calculates the test statistics for continuous data
}
\keyword{internal}
