% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TS_cont}
\alias{TS_cont}
\title{find test statistics for continuous data}
\usage{
TS_cont(x, y)
}
\arguments{
\item{x}{first continuous data set}

\item{y}{second continuous data set}
}
\value{
A vector of test statistics
}
\description{
find test statistics for continuous data
}
\keyword{internal}
