% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TS_disc}
\alias{TS_disc}
\title{find test statistics for discrete data}
\usage{
TS_disc(x, y, vals, ADweights = as.numeric(c(2)))
}
\arguments{
\item{x}{integer vector of data set 1}

\item{y}{integer vector of data set 2}

\item{vals}{numeric vector of values of discrete data set}

\item{ADweights}{A vector of weights for AD method}
}
\value{
A vector of test statistics
}
\description{
find test statistics for discrete data
}
\keyword{internal}
