% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TSw_cont}
\alias{TSw_cont}
\title{find test statistics for continuous data with weights}
\usage{
TSw_cont(x, y, wx, wy)
}
\arguments{
\item{x}{first continuous data set}

\item{y}{second continuous data set}

\item{wx}{weights of x}

\item{wy}{weights of y}
}
\value{
A vector of test statistics
}
\description{
find test statistics for continuous data with weights
}
\keyword{internal}
