% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TSw_disc}
\alias{TSw_disc}
\title{Find test statistics for weighted discrete data}
\usage{
TSw_disc(x, y, vals, wx, wy)
}
\arguments{
\item{x}{integer vector of counts}

\item{y}{integer vector of counts}

\item{vals}{A numeric vector with the values of the discrete rv.}

\item{wx}{integer vector of weights}

\item{wy}{integer vector of weights}
}
\value{
A vector with test statistics
}
\description{
Find test statistics for weighted discrete data
}
\keyword{internal}
