% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weights}
\alias{weights}
\title{find weights for several statistics for discrete data}
\usage{
weights(dta)
}
\arguments{
\item{dta}{A list with vectors x, y and vals.}
}
\value{
A vector of weights
}
\description{
find weights for several statistics for discrete data
}
\keyword{internal}
