% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{power_disc}
\alias{power_disc}
\title{Find the power of various tests via permutation.}
\usage{
power_disc(
  rxy,
  TS,
  typeTS,
  TSextra,
  alpha = 0.05,
  samplingmethod = 2L,
  B = as.numeric(c(1000, 2000)),
  xparam = 0,
  yparam = 0
)
}
\arguments{
\item{rxy}{a function that generates x and y data.}

\item{TS}{routine to calculate test statistics for non-chi-square tests}

\item{typeTS}{indicator for type of test statistics}

\item{TSextra}{additional info passed to TS, if necessary}

\item{alpha}{A numeric constant}

\item{samplingmethod}{=2, 1 for independence sampling, 2 for MCMC in discrete data case}

\item{B}{Number of simulation runs.}

\item{xparam}{arguments for r1.}

\item{yparam}{arguments for r2.}
}
\value{
A numeric matrix of powers
}
\description{
Find the power of various tests via permutation.
}
\keyword{internal}
