% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_newtest.R
\name{power_newtest}
\alias{power_newtest}
\title{This function estimates the power of test routines that calculate p value(s)}
\usage{
power_newtest(TS, f, param_alt, TSextra, alpha = 0.05, B = 1000)
}
\arguments{
\item{TS}{routine to calculate test statistics.}

\item{f}{routine that generates data.}

\item{param_alt}{values of parameter under the alternative hypothesis.}

\item{TSextra}{list passed to TS.}

\item{alpha}{=0.05  type I error.}

\item{B}{= 1000 number of simulation runs to estimate the power.}
}
\value{
A matrix of power values
}
\description{
This function estimates the power of test routines that calculate p value(s)
}
