% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TSw_disc}
\alias{TSw_disc}
\title{find test statistics for discrete data}
\usage{
TSw_disc(x, y)
}
\arguments{
\item{x}{vector of weights data set 1}

\item{y}{vector of weights data set 2}
}
\value{
A vector of test statistics
}
\description{
find test statistics for discrete data
}
\keyword{internal}
